refactor: переместить валидацию значения оценки в контроллер, поменять дефолтное значение на 6

+++ src/main/java/ru/yandex/practicum/filmorate/repository/FilmRepository.java
@@ -45,7 +45,7 @@ public interface FilmRepository {
      * @param grade  Оценка фильма
      * @return Фильм с обновленным списком лайков
      */
-    Film putLike(Integer id, Integer userId, Double grade);
+    Film putLike(Integer id, Integer userId, Integer grade);
 
     /**
      * Удаление лайка пользователя
